src/modules/domain/services/__tests__/TicketViewUsecase.test.ts
import TicketViewUsecase from '../TicketViewUsecase';
import { TicketRepository } from '../../repository/TicketRepository';
import { TicketOutputModel } from '../../model/TicketOutputModel';
import { TicketPresenter } from '../../repository/TicketPresenter';

// Create mock implementations of TicketRepository and TicketPresenter
const mockTicketRepository: TicketRepository = {
  getTicketById: jest.fn(),
  getAllTickets: jest.fn(),
  // Mock other methods as needed
};

const mockTicketPresenter: TicketPresenter = {
  fetchTicket: jest.fn(),
  fetchAllTickets: jest.fn(),
  // Mock other methods as needed
};

describe('TicketViewUsecase', () => {
  let usecase: TicketViewUsecase;

  beforeEach(() => {
    usecase = new TicketViewUsecase(mockTicketRepository, mockTicketPresenter);
  });

  it('should view a ticket successfully', async () => {
    // Arrange
    const ticketId = '1';
    const ticketData = {
      id: '1',
      title: 'Sample Ticket',
      description: 'Description',
      // Add other properties as needed
    };
    mockTicketRepository.getTicketById.mockResolvedValue(ticketData);

    // Act
    const result = await usecase.viewTicket(ticketId);

    // Assert
    expect(result).toEqual({
      id: '1',
      title: 'Sample Ticket',
      description: 'Description',
      // Add other properties as needed
    });
    expect(mockTicketPresenter.fetchTicket).toHaveBeenCalledWith(result);
  });

  it('should handle viewing a non-existent ticket', async () => {
    // Arrange
    const ticketId = '2';
    mockTicketRepository.getTicketById.mockResolvedValue(null);

    // Act
    const result = await usecase.viewTicket(ticketId);

    // Assert
    expect(result).toBeNull();
    expect(mockTicketPresenter.fetchTicket).not.toHaveBeenCalled();
  });

  it('should fetch all tickets successfully', async () => {
    // Arrange
    const ticketDataArray = [
      {
        id: '1',
        title: 'Ticket 1',
        description: 'Description 1',
        // Add other properties as needed
      },
      {
        id: '2',
        title: 'Ticket 2',
        description: 'Description 2',
        // Add other properties as needed
      },
    ];
    mockTicketRepository.getAllTickets.mockResolvedValue(ticketDataArray);

    // Act
    const result = await usecase.getAllTickets();

    // Assert
    expect(result).toEqual([
      {
        id: '1',
        title: 'Ticket 1',
        description: 'Description 1',
        // Add other properties as needed
      },
      {
        id: '2',
        title: 'Ticket 2',
        description: 'Description 2',
        // Add other properties as needed
      },
    ]);
    expect(mockTicketPresenter.fetchAllTickets).toHaveBeenCalledWith(result);
  });

  // Add more test cases as needed for error handling and other scenarios
});
